Added phi.h
Added a concentration function to determine the total concentration of
each polymer block (A,B, and C), as well as the total concentration. I
think that there must be a mistake - the concentration of A and B tend
not to be the same. Perhaps there is a problem with the intermediate
step for the delta-function pinning A and B together.